+++ src/solverpy/solver/plugins/trains/__init__.py
@@ -1,4 +1,6 @@
 from .cvc5 import Cvc5Trains, Cvc5TrainsDebug
+from .enigma import EnigmaTrains, EnigmaTrainsDebug
 
-__all__ = [Cvc5Trains, Cvc5TrainsDebug]
+__all__ = [Cvc5Trains, Cvc5TrainsDebug, 
+           EnigmaTrains, EnigmaTrainsDebug]